iothook/iotv7/requirements.txt
django==4.2.2
djangorestframework
markdown==3.4.3
django-filter==23.2
drf-yasg[validation]